dom/html/test/test_bug386728.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=386728
-->
<head>
  <title>Test for Bug 386728</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=386728">Mozilla Bug 386728</a>
<p id="display"></p>
<div id="content">
  <div id="frameContent">
    <div id="edit">This text is editable</div>
    <button id="button_on" onclick="document.getElementById('edit').setAttribute('contenteditable', 'true')"></button>
  </div>
  <iframe id="testIframe"></iframe>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 386728 **/

var frame = document.getElementById("testIframe");

function testContentEditable() {
  frame.style.display = 'block';
  var frameContent = frame.contentDocument.adoptNode(document.getElementById("frameContent"));
  frame.contentDocument.body.appendChild(frameContent);
  frame.contentDocument.getElementById("edit").contentEditable = "true";
  frame.contentDocument.getElementById("edit").contentEditable = "false";
  frame.contentDocument.getElementById("button_on").click();
  is(frame.contentDocument.getElementById("edit").contentEditable, "true");
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(testContentEditable);
addLoadEvent(SimpleTest.finish);

</script>
</pre>
</body>
</html>
